+++ provisioner/master/spec/tenant_worker_request.rb
@@ -31,11 +31,19 @@ end
 data = Hash.new { |h, k| h[k] = Hash.new(&h.default_proc) }
 data['id'] = id
 data['workers'] = num_workers.to_i
-data['resources']['automatortype']['chef-solo']['cookbooks']['format'] = 'archive'
-data['resources']['automatortype']['chef-solo']['cookbooks']['active'] = [
-      {"name" => "reactor", "version" => '2'},
-      {"name" => "hadoop", "version" => '52'}
-    ]
+#data['resources']['automatortypes']['chef-solo']['cookbooks']['format'] = 'archive'
+#data['resources']['automatortypes']['chef-solo']['cookbooks']['active'] = [
+#      {"name" => "hadoop", "version" => '2'}
+#    ]
+#data['resources']['automatortypes']['chef-solo']['roles']['format'] = 'file'
+#data['resources']['automatortypes']['chef-solo']['roles']['active'] = [
+#      {"name" => "testrole.json", "version" => '2'},
+#      {"name" => "anotherrole.json", "version" => '1'}
+#    ]
+#data['resources']['automatortypes']['chef-solo']['data_bags']['format'] = 'archive'
+#data['resources']['automatortypes']['chef-solo']['data_bags']['active'] = [
+#      {"name" => "users", "version" => '2'}
+#    ]
 
 begin
   json = JSON.generate(data)