+++ provisioner/worker/plugins/providers/fog_provider/lib/fog_provider/aws.rb
@@ -45,7 +45,7 @@ class FogProviderAWS < Provider
       server = connection.servers.create(create_server_def)
       # Process results
       @result['result']['providerid'] = server.id.to_s
-      @result['result']['ssh-auth']['user'] = @task['config']['ssh-auth']['user'] || 'root'
+      @result['result']['ssh-auth']['user'] = @task['config']['sshuser'] || 'root'
       @result['result']['ssh-auth']['identityfile'] = @aws_keyfile unless @aws_keyfile.nil?
       @result['status'] = 0
     rescue Exception => e
@@ -103,7 +103,6 @@ class FogProviderAWS < Provider
       log.debug "Server #{server.id} sshd is up"
 
       # Process results
-      @result['result']['ipaddress'] = bootstrap_ip
       @result['ipaddresses'] = {
         'access_v4' => bootstrap_ip,
         'bind_v4' => bind_ip
@@ -127,8 +126,8 @@ class FogProviderAWS < Provider
       log.error 'Timeout waiting for the server to be created'
       @result['stderr'] = 'Timed out waiting for server to be created'
     rescue Net::SSH::AuthenticationFailed => e
-      log.error("SSH Authentication failure for #{providerid}/#{@result['result']['ipaddress']}")
-      @result['stderr'] = "SSH Authentication failure for #{providerid}/#{@result['result']['ipaddress']}: #{e.inspect}"
+      log.error("SSH Authentication failure for #{providerid}/#{bootstrap_ip}")
+      @result['stderr'] = "SSH Authentication failure for #{providerid}/#{bootstrap_ip}: #{e.inspect}"
     rescue Exception => e
       log.error('Unexpected Error Occurred in FogProviderAWS.confirm:' + e.inspect)
       @result['stderr'] = "Unexpected Error Occurred in FogProviderAWS.confirm: #{e.inspect}"