+++ provisioner/worker/plugins/providers/fog_provider/lib/fog_provider/joyent.rb
@@ -45,7 +45,7 @@ class FogProviderJoyent < Provider
       end
       # Process results
       @result['result']['providerid'] = server.id.to_s
-      @result['result']['ssh-auth']['user'] = 'root'
+      @result['result']['ssh-auth']['user'] = @task['config']['sshuser'] || 'root'
       @result['result']['ssh-auth']['identityfile'] = @joyent_keyfile unless @joyent_keyfile.nil?
       @result['status'] = 0
     rescue Exception => e
@@ -86,11 +86,14 @@ class FogProviderJoyent < Provider
       log.debug "Server #{server.name} sshd is up"
 
       # Process results
-      @result['result']['ipaddress'] = bootstrap_ip
+      @result['ipaddresses'] = {
+        'access_v4' => bootstrap_ip,
+        'bind_v4' => bootstrap_ip
+      }
       # Additional checks
       set_credentials(@task['config']['ssh-auth'])
       # Validate connectivity
-      Net::SSH.start(@result['result']['ipaddress'], @task['config']['ssh-auth']['user'], @credentials) do |ssh|
+      Net::SSH.start(bootstrap_ip, @task['config']['ssh-auth']['user'], @credentials) do |ssh|
         # Backwards-compatibility... ssh_exec! takes 2 arguments prior to 0.9.8
         ssho = method(:ssh_exec!)
         if ssho.arity == 2
@@ -109,8 +112,8 @@ class FogProviderJoyent < Provider
       log.error 'Timeout waiting for the server to be created'
       @result['stderr'] = 'Timed out waiting for server to be created'
     rescue Net::SSH::AuthenticationFailed => e
-      log.error("SSH Authentication failure for #{providerid}/#{@result['result']['ipaddress']}")
-      @result['stderr'] = "SSH Authentication failure for #{providerid}/#{@result['result']['ipaddress']}: #{e.inspect}"
+      log.error("SSH Authentication failure for #{providerid}/#{bootstrap_ip}")
+      @result['stderr'] = "SSH Authentication failure for #{providerid}/#{bootstrap_ip}: #{e.inspect}"
     rescue Exception => e
       log.error('Unexpected Error Occurred in FogProviderJoyent.confirm:' + e.inspect)
       @result['stderr'] = "Unexpected Error Occurred in FogProviderJoyent.confirm: #{e.inspect}"