+++ server/src/main/java/com/continuuity/loom/cluster/NodeProperties.java
@@ -20,15 +20,17 @@ import com.continuuity.loom.admin.ServiceAction;
 import com.google.common.base.Objects;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
+import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
 
+import java.util.Map;
 import java.util.Set;
 
 /**
  * Properties of a node.
  */
 public class NodeProperties {
-  private String ipaddress;
+  private Map<String, String> ipaddresses;
   private final String hostname;
   private final int nodenum;
   // this is the name of the hardware type
@@ -38,33 +40,47 @@ public class NodeProperties {
   // TODO: remove flavor, image, sshUser when hardware/image type switches to the objects instead of names
   private final String flavor;
   private final String image;
-  private final String sshUser;
+  private final String sshuser;
   // list of service names
   private final Set<String> services;
   // list of automators that could be used on the node
   private final Set<String> automators;
 
-  private NodeProperties(String hostname, String ipaddress, int nodenum, String hardwaretype, String imagetype,
-                         String flavor, String image, String sshUser, Set<String> automators, Set<String> services) {
-    this.ipaddress = ipaddress;
+  private NodeProperties(String hostname, Map<String, String> ipaddresses, int nodenum, String hardwaretype,
+                         String imagetype, String flavor, String image, String sshUser,
+                         Set<String> automators, Set<String> services) {
+    this.ipaddresses = ImmutableMap.copyOf(ipaddresses);
     this.hostname = hostname;
     this.imagetype = imagetype;
     this.hardwaretype = hardwaretype;
     this.flavor = flavor;
     this.image = image;
     this.nodenum = nodenum;
-    this.sshUser = sshUser;
+    this.sshuser = sshUser == null ? "root" : sshUser;
     this.automators = automators == null ? ImmutableSet.<String>of() : ImmutableSet.copyOf(automators);
     this.services = services == null ? ImmutableSet.<String>of() : ImmutableSet.copyOf(services);
   }
 
   /**
-   * Get the IP address of the node.
+   * Get an immutable mapping of IP type to IP address.
    *
-   * @return IP address of the node.
+   * @return Immutable mapping of IP type to IP address.
    */
-  public String getIpaddress() {
-    return ipaddress;
+  public Map<String, String> getIPAddresses() {
+    return ipaddresses;
+  }
+
+  /**
+   * Get a specific type of IP address of the node.
+   *
+   * @param ipType Type of IP address to get.
+   * @return IP address of the given type of the node, or null if it does not exist.
+   */
+  public String getIPAddress(String ipType) {
+    if (ipaddresses != null) {
+      return ipaddresses.get(ipType);
+    }
+    return null;
   }
 
   /**
@@ -128,35 +144,34 @@ public class NodeProperties {
    * @return User the provisioner will use to ssh in to the node.
    */
   public String getSshUser() {
-    return sshUser;
+    return sshuser;
   }
 
   /**
-   * Get the names of the automators that can perform actions on the node.
+   * Get an immutable set of the names of the automators that can perform actions on the node.
    *
-   * @return Names of the automators that can perform actions on the node.
+   * @return Immutable set of the names of the automators that can perform actions on the node.
    */
   public Set<String> getAutomators() {
     return automators;
   }
 
   /**
-   * Get the names of the services on the node.
+   * Get an immutable set of the names of the services on the node.
    *
-   * @return Names of the services on the node.
+   * @return Immutable set of the names of the services on the node.
    */
   public Set<String> getServices() {
     return services;
   }
 
   /**
-   * Set the IP address of the node.
+   * Set the IP addresses of the node.
    *
-   * @param ipaddress IP address to set.
+   * @param ipaddresses IP addresses to set.
    */
-  public void setIpaddress(String ipaddress) {
-    this.ipaddress = ipaddress;
-    ImmutableMap.builder();
+  public void setIpaddresses(Map<String, String> ipaddresses) {
+    this.ipaddresses = ipaddresses;
   }
 
   /**
@@ -172,7 +187,7 @@ public class NodeProperties {
    * Builder for creating node properties.
    */
   public static class Builder {
-    private String ipaddress;
+    private Map<String, String> ipaddresses = Maps.newHashMap();
     private String hostname;
     private int nodenum;
     private String hardwaretype;
@@ -183,8 +198,13 @@ public class NodeProperties {
     private Set<String> serviceNames;
     private Set<String> automators;
 
-    public Builder setIpaddress(String ipaddress) {
-      this.ipaddress = ipaddress;
+    public Builder addIPAddress(String ipType, String ipaddress) {
+      this.ipaddresses.put(ipType, ipaddress);
+      return this;
+    }
+
+    public Builder setIPAddresses(Map<String, String> ipAddresses) {
+      this.ipaddresses = ipAddresses;
       return this;
     }
 
@@ -248,7 +268,7 @@ public class NodeProperties {
     }
 
     public NodeProperties build() {
-      return new NodeProperties(hostname, ipaddress, nodenum, hardwaretype, imagetype,
+      return new NodeProperties(hostname, ipaddresses, nodenum, hardwaretype, imagetype,
                                 flavor, image, sshUser, automators, serviceNames);
     }
   }
@@ -265,34 +285,34 @@ public class NodeProperties {
     NodeProperties that = (NodeProperties) o;
 
     return Objects.equal(hostname, that.hostname) &&
-      Objects.equal(ipaddress, that.ipaddress) &&
+      Objects.equal(ipaddresses, that.ipaddresses) &&
       Objects.equal(nodenum, that.nodenum) &&
       Objects.equal(hardwaretype, that.hardwaretype) &&
       Objects.equal(imagetype, that.imagetype) &&
       Objects.equal(flavor, that.flavor) &&
       Objects.equal(image, that.image) &&
-      Objects.equal(sshUser, that.sshUser) &&
+      Objects.equal(sshuser, that.sshuser) &&
       Objects.equal(services, that.services) &&
       Objects.equal(automators, that.automators);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hashCode(hostname, ipaddress, nodenum, hardwaretype, imagetype,
-                            flavor, image, sshUser, services, automators);
+    return Objects.hashCode(hostname, ipaddresses, nodenum, hardwaretype, imagetype,
+                            flavor, image, sshuser, services, automators);
   }
 
   @Override
   public String toString() {
     return Objects.toStringHelper(this)
       .add("hostname", hostname)
-      .add("ipaddress", ipaddress)
+      .add("ipaddresses", ipaddresses)
       .add("nodenum", nodenum)
       .add("hardwaretype", hardwaretype)
       .add("imagetype", imagetype)
       .add("flavor", flavor)
       .add("image", image)
-      .add("sshUser", sshUser)
+      .add("sshuser", sshuser)
       .add("services", services)
       .add("automators", automators)
       .toString();