+++ server/src/main/java/com/continuuity/loom/codec/json/current/TaskConfigCodec.java
@@ -44,9 +44,9 @@ public class TaskConfigCodec extends AbstractCodec<TaskConfig> {
     jsonObj.add("cluster", taskConfig.getClusterConfig());
     jsonObj.add("service", context.serialize(taskConfig.getTaskServiceAction(), TaskServiceAction.class));
     jsonObj.addProperty("hostname", taskConfig.getNodeProperties().getHostname());
-    jsonObj.addProperty("ipaddress", taskConfig.getNodeProperties().getIpaddress());
+    jsonObj.add("ipaddresses", context.serialize(taskConfig.getNodeProperties().getIPAddresses()));
     jsonObj.addProperty("nodenum", taskConfig.getNodeProperties().getNodenum());
-    jsonObj.addProperty("ssh-user", taskConfig.getNodeProperties().getSshUser());
+    jsonObj.addProperty("sshuser", taskConfig.getNodeProperties().getSshUser());
     // these should be changed to be the full object instead of the name
     jsonObj.add("hardwaretype", context.serialize(taskConfig.getNodeProperties().getHardwaretype(), String.class));
     jsonObj.add("imagetype", context.serialize(taskConfig.getNodeProperties().getImagetype(), String.class));
@@ -82,12 +82,13 @@ public class TaskConfigCodec extends AbstractCodec<TaskConfig> {
     TaskServiceAction taskServiceAction = context.deserialize(jsonObj.remove("service"), TaskServiceAction.class);
     // build node properties
     String hostname = context.deserialize(jsonObj.remove("hostname"), String.class);
-    String ipaddress = context.deserialize(jsonObj.remove("ipaddress"), String.class);
+    Map<String, String> ipaddresses = context.deserialize(jsonObj.remove("ipaddresses"),
+                                                          new TypeToken<Map<String, String>>() {}.getType());
     String imagetype = context.deserialize(jsonObj.remove("imagetype"), String.class);
     String hardwaretype = context.deserialize(jsonObj.remove("hardwaretype"), String.class);
     String flavor = context.deserialize(jsonObj.remove("flavor"), String.class);
     String image = context.deserialize(jsonObj.remove("image"), String.class);
-    String sshUser = context.deserialize(jsonObj.remove("ssh-user"), String.class);
+    String sshUser = context.deserialize(jsonObj.remove("sshuser"), String.class);
     Integer nodeNum = context.deserialize(jsonObj.remove("nodenum"), Integer.class);
     Set<String> serviceNames =
       context.deserialize(jsonObj.remove("services"), new TypeToken<Set<String>>() {}.getType());
@@ -95,7 +96,7 @@ public class TaskConfigCodec extends AbstractCodec<TaskConfig> {
       context.deserialize(jsonObj.remove("automators"), new TypeToken<Set<String>>() {}.getType());
     NodeProperties nodeProperties = NodeProperties.builder()
       .setHostname(hostname)
-      .setIpaddress(ipaddress)
+      .setIPAddresses(ipaddresses)
       .setNodenum(nodeNum)
       .setHardwaretype(hardwaretype)
       .setImagetype(imagetype)