+++ server/src/test/java/com/continuuity/loom/scheduler/JobPlannerTest.java
@@ -102,9 +102,9 @@ public class JobPlannerTest {
                                 ProvisionerAction.INITIALIZE, CHEF_ACTION,
                                 ProvisionerAction.START, CHEF_ACTION));
     Node node1 = new Node("node1", "1", ImmutableSet.<Service>of(s1),
-                          NodeProperties.builder().setHostname("host1").setIpaddress("ip1").build());
+                          NodeProperties.builder().setHostname("host1").addIPAddress("access_v4", "ip1").build());
     Node node2 = new Node("node2", "1", ImmutableSet.<Service>of(s1, s2, s3),
-                          NodeProperties.builder().setHostname("host2").setIpaddress("ip2").build());
+                          NodeProperties.builder().setHostname("host2").addIPAddress("access_v4", "ip2").build());
 
     Set<Node> clusterNodes = ImmutableSet.of(node1, node2);
 
@@ -181,9 +181,9 @@ public class JobPlannerTest {
                                 ProvisionerAction.INITIALIZE, CHEF_ACTION,
                                 ProvisionerAction.START, CHEF_ACTION));
     Node node1 = new Node("node1", "1", ImmutableSet.<Service>of(s1),
-                          NodeProperties.builder().setHostname("host1").setIpaddress("ip1").build());
+                          NodeProperties.builder().setHostname("host1").addIPAddress("access_v4", "ip1").build());
     Node node2 = new Node("node2", "1", ImmutableSet.<Service>of(s1, s2, s3),
-                          NodeProperties.builder().setHostname("host2").setIpaddress("ip2").build());
+                          NodeProperties.builder().setHostname("host2").addIPAddress("access_v4", "ip2").build());
 
     Set<Node> clusterNodes = ImmutableSet.of(node1, node2);
 
@@ -276,9 +276,9 @@ public class JobPlannerTest {
                                 ProvisionerAction.START, CHEF_ACTION,
                                 ProvisionerAction.STOP, CHEF_ACTION));
     Node node1 = new Node("node1", "1", ImmutableSet.<Service>of(s1),
-                          NodeProperties.builder().setHostname("host1").setIpaddress("ip1").build());
+                          NodeProperties.builder().setHostname("host1").addIPAddress("access_v4", "ip1").build());
     Node node2 = new Node("node2", "1", ImmutableSet.<Service>of(s1),
-                          NodeProperties.builder().setHostname("host2").setIpaddress("ip2").build());
+                          NodeProperties.builder().setHostname("host2").addIPAddress("access_v4", "ip2").build());
 
     Set<Node> clusterNodes = ImmutableSet.of(node1, node2);
 
@@ -312,7 +312,7 @@ public class JobPlannerTest {
 
     // add services s2 and s3 to node2
     node2 = new Node("node2", "1", ImmutableSet.<Service>of(s1, s2, s3),
-                     NodeProperties.builder().setHostname("host2").setIpaddress("ip2").build());
+                     NodeProperties.builder().setHostname("host2").addIPAddress("access_v4", "ip2").build());
     clusterNodes = ImmutableSet.of(node1, node2);
     job = new ClusterJob(JobId.fromString("123-002"), ClusterAction.ADD_SERVICES,
                          ImmutableSet.<String>of(s2.getName(), s3.getName()), ImmutableSet.<String>of(node2.getId()));
@@ -417,7 +417,7 @@ public class JobPlannerTest {
                                 ProvisionerAction.START, CHEF_ACTION,
                                 ProvisionerAction.STOP, CHEF_ACTION));
     Node node1 = new Node("node1", "1", ImmutableSet.<Service>of(s1, s2, s3),
-                          NodeProperties.builder().setHostname("host1").setIpaddress("ip1").build());
+                          NodeProperties.builder().setHostname("host1").addIPAddress("access_v4", "ip1").build());
 
     Set<Node> clusterNodes = ImmutableSet.of(node1);
 
@@ -501,7 +501,7 @@ public class JobPlannerTest {
                                 ProvisionerAction.START, CHEF_ACTION,
                                 ProvisionerAction.STOP, CHEF_ACTION));
     Node node1 = new Node("node1", "1", ImmutableSet.<Service>of(s1, s2, s3),
-                          NodeProperties.builder().setHostname("host1").setIpaddress("ip1").build());
+                          NodeProperties.builder().setHostname("host1").addIPAddress("access_v4", "ip1").build());
 
     Set<Node> clusterNodes = ImmutableSet.of(node1);
 
@@ -583,9 +583,9 @@ public class JobPlannerTest {
                               ImmutableMap.<ProvisionerAction, ServiceAction>of(
                                 ProvisionerAction.CONFIGURE, CHEF_ACTION));
     Node node1 = new Node("node1", "1", ImmutableSet.<Service>of(s1),
-                          NodeProperties.builder().setHostname("host1").setIpaddress("ip1").build());
+                          NodeProperties.builder().setHostname("host1").addIPAddress("access_v4", "ip1").build());
     Node node2 = new Node("node2", "1", ImmutableSet.<Service>of(s1, s2, s3),
-                          NodeProperties.builder().setHostname("host2").setIpaddress("ip2").build());
+                          NodeProperties.builder().setHostname("host2").addIPAddress("access_v4", "ip2").build());
 
     Set<Node> clusterNodes = ImmutableSet.of(node1, node2);
 
@@ -622,9 +622,9 @@ public class JobPlannerTest {
                                 ProvisionerAction.START, CHEF_ACTION,
                                 ProvisionerAction.STOP, CHEF_ACTION));
     Node node1 = new Node("node1", "1", ImmutableSet.<Service>of(s1),
-                          NodeProperties.builder().setHostname("host1").setIpaddress("ip1").build());
+                          NodeProperties.builder().setHostname("host1").addIPAddress("access_v4", "ip1").build());
     Node node2 = new Node("node2", "1", ImmutableSet.<Service>of(s1, s2, s3),
-                          NodeProperties.builder().setHostname("host2").setIpaddress("ip2").build());
+                          NodeProperties.builder().setHostname("host2").addIPAddress("access_v4", "ip2").build());
 
     Set<Node> clusterNodes = ImmutableSet.of(node1, node2);
 
@@ -693,9 +693,9 @@ public class JobPlannerTest {
                                 ProvisionerAction.START, CHEF_ACTION,
                                 ProvisionerAction.STOP, CHEF_ACTION));
     Node node1 = new Node("node1", "1", ImmutableSet.<Service>of(s1),
-                          NodeProperties.builder().setHostname("host1").setIpaddress("ip1").build());
+                          NodeProperties.builder().setHostname("host1").addIPAddress("access_v4", "ip1").build());
     Node node2 = new Node("node2", "1", ImmutableSet.<Service>of(s1, s2, s3),
-                          NodeProperties.builder().setHostname("host2").setIpaddress("ip2").build());
+                          NodeProperties.builder().setHostname("host2").addIPAddress("access_v4", "ip2").build());
 
     Set<Node> clusterNodes = ImmutableSet.of(node1, node2);
 
@@ -756,9 +756,9 @@ public class JobPlannerTest {
                                 ProvisionerAction.START, CHEF_ACTION,
                                 ProvisionerAction.STOP, CHEF_ACTION));
     Node node1 = new Node("node1", "1", ImmutableSet.<Service>of(s1),
-                          NodeProperties.builder().setHostname("host1").setIpaddress("ip1").build());
+                          NodeProperties.builder().setHostname("host1").addIPAddress("access_v4", "ip1").build());
     Node node2 = new Node("node2", "1", ImmutableSet.<Service>of(s1, s2, s3),
-                          NodeProperties.builder().setHostname("host2").setIpaddress("ip2").build());
+                          NodeProperties.builder().setHostname("host2").addIPAddress("access_v4", "ip2").build());
 
     Set<Node> clusterNodes = ImmutableSet.of(node1, node2);
 
@@ -821,9 +821,9 @@ public class JobPlannerTest {
                                 ProvisionerAction.START, CHEF_ACTION,
                                 ProvisionerAction.STOP, CHEF_ACTION));
     Node node1 = new Node("node1", "1", ImmutableSet.<Service>of(s1),
-                          NodeProperties.builder().setHostname("host1").setIpaddress("ip1").build());
+                          NodeProperties.builder().setHostname("host1").addIPAddress("access_v4", "ip1").build());
     Node node2 = new Node("node2", "1", ImmutableSet.<Service>of(s1, s2, s3),
-                          NodeProperties.builder().setHostname("host2").setIpaddress("ip2").build());
+                          NodeProperties.builder().setHostname("host2").addIPAddress("access_v4", "ip2").build());
 
     Set<Node> clusterNodes = ImmutableSet.of(node1, node2);
 
@@ -884,9 +884,9 @@ public class JobPlannerTest {
   public void testNoEdgeNodesInDag() {
     Service s1 = new Service("s1", "", ImmutableSet.<String>of(), ImmutableMap.<ProvisionerAction, ServiceAction>of());
     Node node1 = new Node("node1", "1", ImmutableSet.<Service>of(s1),
-                          NodeProperties.builder().setHostname("host1").setIpaddress("ip1").build());
+                          NodeProperties.builder().setHostname("host1").addIPAddress("access_v4", "ip1").build());
     Node node2 = new Node("node2", "1", ImmutableSet.<Service>of(s1),
-                          NodeProperties.builder().setHostname("host2").setIpaddress("ip2").build());
+                          NodeProperties.builder().setHostname("host2").addIPAddress("access_v4", "ip2").build());
     Set<Node> clusterNodes = ImmutableSet.of(node1, node2);
 
     TaskDag expected = new TaskDag();